style: adds linting config and uses it
Linting can be applied by running the following command:

    npm run lint

This is taken from the Node.js standard, with some modifications
to follow the style already present in priority-queue.js. Ideally,
we would be referring to centralised Node.js style guide, and any
disagreements could be resolved with a pull request to that guide.

I look forward to the fiery 	 vs      debates!