refactor: change PriorityQueue to a class
Using the 'class' concept here makes the code much more readable and
modern, and doesn't outwardly change the behaviour of any of our happy
path functionality (we can see the proof of this since our tests at this
revision still pass).

Using a class also gives us all of the advantages of explicitly
specifying 'prototype' like you did for free. Incidentally, this moves
the 'add,' 'Pop,' and 'length' methods onto the prototype of this class,
with no significant difference to usage.