feat: make class iterable
While here, deprecate our forEach implementation. I don't recommend
using forEach with modern Node.js. It's slower than a for loop, and
can break async/await behaviour if you're not careful.

Technically, `function*` or `*foo` or `*[Symbol.iterator()]` in our
case, returns a generator, which is a function that conforms to the
rules of an iterable (like how an Array or a Map are iterable).