fix: throw error if adding item with bad priority
Enforcing priority type here allows us to side-step a bunch of nasty
behaviour and unnecessary type checking. It's also a good illustration
of a problem that wouldn't exist if we used TypeScript.

BREAKING CHANGE: If we're trying to follow SOLID here, and you consider
our PriorityQueue a subtype of Map (I don't, but if you did...), then
throwing this TypeError would break LSP. Food for thought. :)