fix: limit supported datatypes of our queue
This makes a whole load of complexity just go away. Maybe if we had
gathered the use case(s) for this module we wouldn't have to take such
drastic measures, and we could target our implementation to solve a real
world problem.

Alas, we didn't take any such precautions, and so we now take the
drastic measures you see in this patch.